+++ pages/home.tsx
@@ -5,6 +5,7 @@ import { AdminHomePage } from "../components/HomePage/AdminHomePage";
 import { UserHomePage } from "../components/HomePage/UserHomePage";
 import { Error } from "../components/util/Error";
 import styles from "../styles/Home.module.css";
+import Link from "next/link";
 
 //This is the page that is rendered when the 'Home' button from the Navbar is clicked
 const Home: NextApplicationPage = () => {
@@ -14,6 +15,10 @@ const Home: NextApplicationPage = () => {
 
   return (
     <div>
+
+      <Link href="/class/1">
+        <a>Class</a>
+      </Link>
       {user.role == "Admin" ? <AdminHomePage /> : <UserHomePage userId={user.id} />}
       <div className={styles.spacing} />
     </div>
